(PUP-7104) Fix problem with nil/default value in Object serializer
The object serializer did not serialize nil values if they were optional.
This was incorrect since the default value for an attribute may be something
other than `nil`. This commit fixes the problem, and then also fixes some
incorrect type definitions that already existing tests discovered when the
problem was fixed.