(MAINT) Skip puppet_lookup_cmd.rb test when no non-master agents present
Previously, if the host configuration for a `puppet_lookup_cmd.rb` test
run were to include a single node which has both the 'master' and 'agent'
role, the 'host' would be set to nil, causing the test to fail as it hit
the first line of code which tried to make calls on the 'host' object.

This commit causes the `puppet_lookup_cmd.rb` test to be skipped if no
agent nodes that do not have the master role are present.