(PUP-6921) Improve separation of concern between String and Enum types
Before this commit, the `PStringType` had two attributes. A `size_type` and
a `values` array. An inferred String would hence result in a String type
with two complex objects and all checkes would then take both objects into
account.

This commit simplifies the `PStringType` so that it only has one attribute.
The attribute can be either a size_type or a string value.

Since this means that a String can hold max one string value, the common
type of two different strings is now changed to instead be an Enum. I.e.
the inferred element type of the array ['a', 'b']. The logic for computing
assignability and commonality between Enum and String has been improved.