(maint) Ensure that parent is a PObjectType when traversing equality
Prior to this commit, an Object type parented by a type that was an Alias
or a type that didn't resolve to an Object type would call a non existing
method `equality_attributes` in case an equality attribute was defined as
a duplicate. It worked for TypeAlias since passes no-response calls to
it's contained type but it might break mysteriously in future additions
to the code.

This commit ensures that the parent traversal uses resolved Object types
only.

(cherry picked from commit 3bb8927)