(PUP-6926) Enable lookup of rich data values
This commit changes the Pcore type used to validate data found by a lookup
so that it may contain any type except Runtime. The key and value types are
registered as aliases `Puppet::LookupKey` and `Puppet::LookupValue`. The
latter alias is self-recursive in that keys and values found in hashes and
arrays must conform to their respective type, even if nested.