(PUP-6188) consider umask for default perms FS test
Prior to this commit, a test in file_system_spec.rb would assert an expected
default permission for a newly created file by Puppet::FileSystem, but failed
to take into account that this would vary based on umask (which varies in linux
by root vs non-root user). This commit updates the test to instead calculate
the expected permission based on the current umask.

Because ruby mode/umask does not behave the same way on Windows, we special
case this behavior to supply what we would expect on Windows.

Paired with: Ethan Brown <ethan@puppet.com> (@iristyle)