(PUP-4742) Add runtime dependency on semantic_puppet
Previously, when puppet was built as a gem, e.g. bundle exec rake
package:gem, it was missing the semantic_puppet runtime dependency.

This commit updates project_data.yaml so that the puppet gem expresses
its runtime dependency on semantic_puppet with the same constraints that
are specified in Gemfile and .gemspec.