(PUP-6967) Reload puppetserver using HUP
Bump beaker to 3.5.x, which added support for reloading puppetserver
using HUP.

Beaker will fail to reload the service if it isn't already running and
it will fall back to bouncing the service using puppet's resource
service ensure=stopped and ensure=running commands. To ensure reloading
always works, we now start puppetserver in the 045_EnsureMasterStarted
pre-suite step. We don't test against passenger or webrick, so there's
no point in making the step conditional on the server type.

Prior to commit 59cce2311, Beaker's `with_puppet_running_on` method would
start puppetserver twice. Once in the beginning before the test, and
once after. Each start takes ~30 seconds per test, and that caused an
overall 1+ hour increase in test times as compared to when we tested
against passenger packages.

In commit 59cce2311, we set `restart_when_done` to true to avoid the
second start, which saved ~30 minutes. However, we have to undo that
change in order to use beaker's reload functionality.

That said, we don't regress on test times, because reloading avoids both
puppetserver starts, and puts us back on par with test times prior to
the move to puppetserver.

This commit also moves the logic for setting `use-service` from an
explicit pre-suite step to the options hash, where all of the other
options are specified. Beaker converts the symbol to a string when
processing the options, so the existing behavior is preserved.