(PUP-6850) Exclude commits unrelated to a PR
TRAVIS_COMMIT_RANGE uses triple dots, which includes commits that are in
the base branch that are not in the PR. For example, if a commit is
pushed to master with an incorrect summary, then every PR created before
that commit will include the unrelated commit, causing the PR to fail
the summary check.

As suggested in https://github.com/travis-ci/travis-ci/issues/4596#issuecomment-139811122,
replace triple dots with double dots.