(PUP-6973) Make hiera configuration sensitive to scope changes
When a hiera configuration contains values with interpolated expressions
that reference scope variables, there is a potential that such values may
change. A variable can be unset on first lookup and introduced later on, or
it may be a global variable shadowed by a local variable (or vice versa).

This commit ensures that such changes are detected and that the
configuration is reloaded.