(maint) Use Windows line endings for batch file in integration test
Previously the execution_spec.rb test was using LF line endings during the UTF8
test which worked on US English systems.  However when running on a Japanese
version of Windows, this test was failing incorrectly.  This commit forces the
line endings to Windows style CRLF so that the batch file will execute correctly.