(PUP-6980) Reroute Global Hiera v3 calls to new lookup framework
This commit changes the global lookup. Instead of directly using the
`data_binding_terminus`, it first checks if it is set to 'hiera' and if that
is the case, it reroutes the lookup to the new lookup framework.

The lookup framework is changed to also handle v3 configurations and a
GlobalDataProvider is added that ensures that lookups that uses the a v3
config also is passed the special Hiera scope that contains the scope
variables 'calling_class', 'calling_class_path', 'calling_module', and
'module_name'. This data provider also ensures that merge options declared
in the hiera.yaml file are used during merge of data from the global
backends when no other options are passed down.

In order to differntiate between the 'first' found merge strategy (which is
the default) and the fact that no merge strategy has been given, a new
'default' merge strategy is added. A reverse 'deep merge' strategy is also
added to deal with the Hiera 3 'deep' (as opposed to 'deeper') merge
behavior.