(PUP-6986) Sanitize ps output when searching for pids
This patch updates the base service provider such that ps output is
sanitized to clean "UTF-8" when a process table scan is run by
service resources that have `hasstatus => false` set. This scan runs
regex matches against the output of `ps -ef` which could contain
arbitrary byte sequences depending on which services are running on the
system and how they are configured. Byte sequences that don't match
UTF-8 will be replaced with U+FFFD characters.