(PUP-6777) handle varying encodings from Etc in nameservice provider
The nameservice provider leverages the ruby Etc::getpwnam method to return
information regarding users (subclasses include useradd and pw). In Ruby <
2.1.0, this method always returns values as strings in binary encoding. In Ruby
2.1.0 and later, this method returns strings in either the current locale
encoding (Encoding::default_external) OR binary encoding if the string returned
does not encode correctly in the current locale encoding (how they determine
this is opaque, at best, and not exposed publicly).

Prior to this commit, the :comment property of the user type would set the
external encoding of the supplied value for :comment to binary if running ruby
< 2.1.0. This effectively handles the case of incompatible encodings with ruby
< 2.1.0. However, for ruby >= 2.1.0, this will always fail in two scenarios:

- if the value of :comment contains character(s) that are not represented in
  the current locale, because ruby Etc will supply this 'is' value in binary and
  we will compare it to our 'should' value in whatever the current locale is,
  failing.

- if the current locale results in a system (is) value of :comment in a ruby
  Encoding incompatible with the formatting of the desired 'should' value from
  Puppet.

This commit instead overrides the Puppet::Property#insync? method for
Puppet::Type::User::Comment when using the nameservice provider, using
`force_encoding` to set the External encoding of the 'should' value to that of
the system 'is' value. This commit also overrides Puppet::Property#change_to_s
for Puppet::Type::User::Comment for scenarios where we are about to concatenate
strings with incompatible character encodings. In this case as well, set the
encoding of the new value to the encoding of the value returned by the system
for printing a message to the user. In both cases, we avoid manipulating the
external encoding of the underlying 'should' object, for fear of unexpected
side-effects in Puppet. One such side-effect was found in
Puppet::Util::Yaml::Dump, which will fail to write the run report if the
should value cannot actually be represented by it's assigned default external
encoding (i.e., if `valid_encoding?` would return false).

Signed-off-by: Moses Mendoza <moses@puppet.com>