(maint) Skip broken test
The acceptance test creates the environment directory with permissions
666 and root:root user/group. As a result, the puppet master running as
`puppet` cannot traverse the directory to read site.pp. For some reason
puppetserver ignores the error, so the test passes in normal CI.

When the directory permissions are changed to 755, the puppet master
fails to parse site.pp, because the data contains quotes that aren't
escaped. When the quotes are fixed, the puppet master fails to parse
site.pp, because it references a `merge` function which isn't present in
puppet core. There is one in stdlib, but we don't install that module
during puppet acceptance.

This commit fixes the permission and quote problem, but skips the test
until PUP-6994 can be resolved.