(PUP-6980) Split test suite for lookup function into two files
This commit splits the long test suite spec/unit/function/lookup_spec into
two files. One that is based on a fixture in the file system and one that
is based on fixtures generated on the fly by the test itself.

Prior to this, the suite was split into two test contexts with fairly long
and somewhat unmotivated test names as the result.

The commit does not alter, add or remove any of the tests.