(pe-19366) Gather resource (package) inventory
Here we add the ability to gather package inventory information and add
it to the report.

This functionality defaults to being disabled, and can be enabled by
setting the `pe_enable_resource_inventory` configuration option to `true`.
When enabled Puppet::Indirector.search will be used to gather a list of
all package resources and add it to a new "resource_inventory" field in
the report.

While resource gathering is currently limited to pacakges, the intent is
to add more resources in the future.