(PUP-6996) Add correct permissions to default production environment
When Puppet is run, it will create the production environment if it
is not already present.  If that run is done at the command line,
the resulting directory will be owned by root:root.  This is not
consistent with the ownership required by Code Manager and file
sync.

This patch corrects the permissions to be those set in the Puppet
configuration.  This prevents any possible permissions issues with
other Puppet components.  If Puppet is running as non-root or
as a user different than the configured Puppet user, we fail back
to the default permissions.