(PUP-6474) Exclude problematic files from gettext parsing
This commit brings in a new version of the gettext setup gem that allows
the user to specify a list of files to be skipped when searching for
translatable strings. This was necessary because some files in the
puppet type system use underscores in a way that triggers the gettext
parser incorrectly. So we need to ignore these files when extracting
string into the POT file.