(PUP-6989) Interpolate GEM_SOURCE on the coordinator
Previously, we created a Gemfile on the remote system with
ENV['GEM_SOURCE'] as the default source. But the environment variable is
never set on the remote system, instead it's typically set on the
coordinator when running beaker.

This commit interpolates the GEM_SOURCE on the coordinator so that all
of the remote systems use that value when installing gems.