(PUP-6989) Update for Solaris and Windows
Create a temp directory on the remote system, and git clone into that
directory. Note the temp directory by default is not readable by
non-root users, which is a problem when puppet master switches users to
`puppet`. So chmod 755 to ensure we don't have permissions problems
loading code later.

On Solaris, we install ruby 1.9, so restrict json_pure to 1.8.3. We also
have to symlink the `bundle` executable like we do for `gem`, etc.

Bundler doesn't support binstubs on Windows[1], and we need to update
our packer Windows images to include Git on Windows, instead of using
Cygwin git. So for now raise an ArgumentError when trying to use `git`
acceptance on Windows hosts.

[1] https://github.com/bundler/bundler/issues/4512