(PUP-6981) Reroute hiera_xxx functions to new lookup framework
This commit changes the hiera_xxx functions so that they use the new lookup
framework instead of being hard-wired directly to the Hiera v3 code.

The functions will only use the global layer unless a hiera.yaml version 5
is present in the environment.

A special `data_hash` function dispatcher was designed to ensure that any
"override paths" that the hiera_xxx functions may provide are inserted
correctly into the hierarchy.