(PUP-6538) Deprecate hiera_xxx functions
This commit adds a deprecation text to the documentation of the four
functions `hiera`, `hiera_array`, `hiera_hash`, and `hiera_include` that
informs the user that `lookup`should be used instead. Code is also added to
ensure that deprecation messages are logged when the functions are used.
The deprecation logging can be silenced by setting --strict setting to off.