(PUP-6982) Allow keys to be specified using patterns in lookup_options
Before this commit, in order for a key to have lookup options specified,
the options had to be keyed with that same exact key. Now, such options can
also be defined for keys that matches a specific pattern. This is achieved
by treating all keys that start with a '^' as regular expression patterns.
A key that matches such a pattern is considered to have the options below
that key.

The commit also adds a check that key or pattern under `lookup_options`
in a module only can identify keys that belong to that module. Keys must
 start with "<module_name>::" and patterns must start with "^" followed
 by "<module_name>::".