(PUP-6981) Fix issue with recursive guard detection when validating types
Lookup introduces a new recursive type `Puppet::LookupValue` which defines
all values that are legal for lookup (in essence, everything that can be
stored in a catalog). The type is recursive since hashes and arrays can
values of the same type.

The acceptance test for hiera_hash, uses a hash with two values that are
exactly the same (uid and gid are both 500). This revealed a bug in the
guard used for trapping endless recursion in recursive types where values
that are pushed didn't get popped after the branch where the recursiveness
was tested. This commit fixes that bug.