(PUP-869) Make it easy to deprecate faces
PUP-869 tracks the deprecation of the `file` face - one of many upcoming
deprecations of existing faces in the puppet codebase. After initial review,
issuing a simple deprecation warning in a face proved more complex than one
might hope - especially in testing that face is actually deprecated when a face
is actually used.  I.e., we could just put a deprecation warning in the initial
face definition block, but this block is only evaluated once, at load time - a
deprecation there might be fine from a ux perspective but is executed via an
`instance_eval` and difficult to test. This commit adds a simple `deprecated`
attribute to the Puppet::Interface class, and ensures that if a face is
deprecated, we issue an appropriate deprecation warning when an action is sent
to that face.

The (perhaps most appreciable) benefit of this is that it allows us to
deprecate a face with a simple call to `face.deprecate` in that face's
definition block.