(PUP-6981) Ensure that legacy Hiera logger is initialized
Before this commit, the class Hiera wasn't loaded properly and its logger
was not initialized. Consequently, any custom backend that attempted to use
the logger would run into an error similar to: "Unable to instantiate
backend '<backend name>': undefined method 'debug' for nil:NilClass".

This commit adds the proper require for the class and ensures that the
logger is initialized.