(PUP-6992) Fix inconsistent autoload of environment puppet functions
This commit changes the autoload behavior for puppet functions and types
that reside under the 'functions' or 'types' directory in an environment
so that the naming becomes consistent with the ruby functions that
reside under 'lib/puppet/functions'.

In essence, the namespace 'environment' is moved into a subdirectory with
the same name that functions and types in the global namespace can reside
directly under their respective 'functions' and 'types' directory.