(PUP-6134) Explicitly locate gemcmd windows
Puppet is unable to install ruby gems into anything other than Puppet's ruby on Windows. In Windows, the puppet ruby bin directory is prepended to the PATH (in environment.bat) and puppet uses the which method in util.rb to find the first gem command in the PATH.

This fix updates gem.rb to use a custom which method to find the first non-puppet gem command in the PATH. This fix also updates puppet_gem.rb to remove its dependency on the puppet ruby bin directory being prepended to the PATH.

Alternate solutions include: implement the optional 'path' attribute of exec resources in providers (easier) and not prepending the puppet ruby bin directory to the PATH in Windows (harder).