Partial Revert "(PUP-6659) Remove Stacktraces in API responses"
This partially reverts commit 7a765b70bb392f1d6ce9c230686f780d249720d0.

Commit 7a765b7 both removed the stack trace array from the http error object
and introduced the stack trace to the puppet error log. We revert the removal
of the stack trace from the http error object because it was a
backwards-incompatible change. However, we retain the addition of the stack
trace to the puppet error log, to avoid surprising anyone who had already begun
to rely on this behavior and because in a subsequent commit we will be
replacing the stack trace string content in the http error object with a
deprecation warning - retaining the content for the logs will allow users to
still find this information if they need it.

Signed-off-by: Moses Mendoza <moses@puppet.com>