(PUP-5756) remove dependency on win32/eventlog
This commit removes our dependency on eventlog by including our own ffi
definitions of the required functions for puppet logging. This is a simple
implementation - write-only support for event log.

This commit is API-compatible with the previous implementation.

Our new eventlog API uses W (unicode) versions of API functions. The spec for
these functions can be found here:
https://msdn.microsoft.com/en-us/library/windows/desktop/aa363654(v=vs.85).aspx

We replace existing log destination of win32/eventlog with
Puppet::Util::Windows::EventLog. This means we don't need to check for the
eventlog library any more as part of the eventlog feature. We keep the eventlog
feature for backwards compatibility for now, but it can be removed.

Note that the previous implementation with Win32::EventLog actually
(unnecessarily) created two handles to the event log - one opened with source
'Application' which we never wrote to, and a second opened with source
'Puppet', which we used. Now we just create one handle, on initialization, with
event source 'Puppet' (hence no longer needing to supply this as a param in
report_event). This will still use the 'Application' log and report from a
source of 'Puppet'.

In order to fully remove our dependency on win32-eventlog, we need to cease
using it in ext/windows/service/daemon.rb, which containst the code that runs
Puppet as a service on Windows. As such, puppet/util/windows/eventlog is
written so that it can be used completely independently of Puppet, and
ext/windows/service/daemon.rb is updated to leverage it.  This comes at the
cost of some duplication - especially with respect to constants and FFI
utilities defined elsewhere. However, this is offset by allowing us to
centralize all of our event reporting code in one place, instead of duplicating
this in eventlog.rb.

Since we no longer use the win32-eventlog gem anywhere, this commit removes it
from `ext/project_data`.

Signed-off-by: Moses Mendoza <moses@puppet.com>