(PUP-7061) Fix problem with unbalanced pop of debug explainer
The debug explainer that was added to an invocation and that emitted
debug info on each resolution of an interpolation, was incorrectly
removed before the interpolation call was finished. This resulted in
an attempt to do a pop on an attribute that no longer existed.

This commit ensures that the attribute is treated as read-only and
instead moves the ability to do normal explaining to the debug by having
it delegate to it's wrapped explainer.