(PUP-5434) Remove deprecated experimental lookup implementation
This commit removes the modules `Puppet::DataProviders` and
`Puppet::Plugins::DataProviders` and all references to the bindings
introduced by the latter.

As a consequence of this removal, it is no longer possible to create special
bindings (under lib/puppet/binidngs) for DataProvider subclasses the way it
was done in the experimental lookup framework. Such providers must be
converted into Hiera version 5 data provider functions.