(PUP-6112) Make trusted_server_facts always be true
This changes trusted_server_facts setting such that it is always as if
it was set to true. The setting itself is deprecated (rather than
removed) to not cause immediate failure, but the settings has no effect.

The default of the setting has changed to true. The acceptance test is
updated to check that it is true, and the test of the false branch is
dropped.