(PUP-869) Add face deprecation message to help summaries
Previously in commit 7fefceae40d5ac a deprecation message was added to the face
and the help text for a face.  However when running `puppet help` it was not
obvious which faces were deprecated or not.

This commit adds the text `(Deprecated)` to the face summary when running
`puppet help`.