(PUP-3534) deprecate Puppet::SSL::CertificateAuthority#list_certificates
This commit deprecates Puppet::SSL::CertificateAuthority#list_certificates, as
it is no longer used by anything. The PE License check code that originally
used this been migrated to querying PuppetDB, and was removed from PE puppet in
the PE 3.1 series.

To avoid superflous warnings emanating from use of
Puppet::SSL::CertificateAuthority#list, that method is updated to directly do
an indirector search instead of indirectly through #list_certificates. This is
what it originally did, before #list_certificates was added, and is
backwards-compatible.