(PUP-3534) tag Puppet::SSL::CertificateAuthority#x509_store as @deprecate
Puppet::SSL::CertificateAuthority#x509_store is a caching layer for the x509
object that was used by the PE License code. This commit tags it as deprecated,
as the PE License code that used it was removed in the PE 3.1 series, and it
was questionably advisable back then to return a potentially stale x509 store
anyway.

We also update #verify to call #create_x509_store instead of #x509_store,
skipping the cache layer and always creating a new object - the behavior before
x509_store was introduced. This is semantically API backwards-compatible, and
now we can remove #x509_store at a major version, as its only callers are also
deprecated.

Note #x509_store is API private, so we don't actually do a
Puppet.deprecation_warning. The comment tag should be enough to warn off anyone
trying to use it before it's removed.