(PUP-7087) Optimize lookup type checking and caching of merge strategy
This commit optimizes the MergeStrategy so that it uses fixed instances
when no options are given (which is true in 99.9% of the cases). It also
changes how the value is validated with respect to type so that it isn't
validated unless a type is given (basic Puppet::LookupValue validation
still takes place in the provider).