(PUP-6930) Modify test for logic that is unaffected by --strict
While the change in PUP-6930 changes the DUPLICATE_DEFAULT from being
under the control of --strict, it was decided that it was not meaningful
to unit test that the --strict flag has no effect on the outcome.

This commit changes the test to remove looping over :off, :warning and
:error. The tests did run green manually before being made into a single
assertion (for the default --strict=warning; which is good enough).