(PUP-7089) Add ability to declare a Hiera 3 backend in a version 5 config
This commit adds the "function type" hiera3_backend. The value of such an
entry must be the name of a Hiera 3 backend (so not a function after all).
When used, a Hiera 3 config will be created on-the-fly and assigned to
the Hiera 3 class. Lookups will then be dispatched to the backend just
as if a Hiera version 3 configuration was used.