(PUP-6958) Remove build-time dependency on the gettext-setup gem
This commit creates wrapper tasks for each of the gettext-setup gem's
rake tasks, to avoid loading the gem automatically every time the
Rakefile is evaluated. This was an issue when packaging the puppet gem,
because the gem builder machine does not have gettext-setup installed.
Since the gem is not actually needed at build time, this removes the
dependency.