(PUP-7095) Enable portage install and uninstall options
Allow the portage package provider to accept install and uninstall
options.  This is useful if you do not want to use the default global
EMERGE_DEFAULT_OPTS paramaters setable in make.conf.

Allow the portage package provider to purge package if requested.

Allow package sets (@world @system and the like)

Use ||= where possible and class variables when appropriate to reduce
runtime.  Old runtime was ~8.1 seconds, new runtime is ~5.4 seconds for
32 packages (and a bunch of other things).