(PUP-4947) Make settings_spec for windows specific test conditional
In settings_spec, there were a test that simulated running on windows by
mocking the windows_feature to true. This used to work but now with app
management on by default, there is more activity in the setup. This in
turn leads to puppet wanting to load things that are not initialized in
the Util::Windows name space.

For that reason, the windows specific test in settings was made
conditional on the windows feature.