(PUP-5954) Change how 'settings' class is set in type collection
Prior to this, a 'settings' class was always added to the
'known_resource_types' collection and this could cause a merge with
previous version of settings. This typically occurs when tests are
executed in sequence with one and the same environment instance.

This is now changed by adding a method 'replace_settings' that either
just adds the 'settings' class or replaces one that was previously
added.