(PUP-7021) Use UTF-8 encoding in Puppet::Provider::NameService#listbyname
The #listbyname method is used in the nameservice provider to generate the
values for self.instances. If a name returned by the Etc module is a utf-8
value but the system encoding is not, self.instances, and thus things like
`puppet resource user` fail with encoding errors.

This commit addresses this by setting the external encoding (not modifying byte
values) of the returned string to UTF8. This is one of the entry points for
these strings into the puppet system, so we prevent a cascade of downstream
failures by setting this to utf-8 here.

Signed-off-by: Moses Mendoza <moses@puppet.com>