(PUP-7021) Use UTF-8 encoding in Puppet::Provider::NameService#groups
The #groups method is used by the nameservice provider (not surprisingly) in
querying group membership. We use the Etc module to parse the group file, which
can return strings in system encoding. Ruby considers two identical stringis
(in bytes) with different external encodings to be unequal. Since the names in
the groups array returned by Etc can be in system encoding, and our user is in
utf-8 encoding, our group membership check will fail.

This commit addresses this by setting the external encoding (not modifying byte
values) of the strings returned by Etc to UTF8. This is one of the entry points
for these strings into the puppet system, so we prevent a cascade of downstream
failures by setting this to utf-8 here.

Signed-off-by: Moses Mendoza <moses@puppet.com>