(PUP-7021) Use UTF-8 encoding in Puppet::Provider::NameService#info2hash
the #info2hash method in the nameservice provider leverages Etc module to
gather information about a user or group. However, the strings returned by the
module can be in system encoding, rather than UTF-8. Later, when puppet tries
to use the information for comparison or display purposes with Puppet's UTF-8
strings, we fail in many, many ways.

This commit addresses this by setting the external encoding (not modifying byte
values) of the strings returned by Etc to UTF8. This is one of the entry points
for these strings into the puppet system, so we prevent a cascade of downstream
failures by setting this to utf-8 here.

Signed-off-by: Moses Mendoza <moses@puppet.com>