(PUP-7042) Fix spec failures
This commit fixes several typos introduced by string marking, as well as
stubs out the `_` translate method in unit tests, where it interferes
with some of our mocking. This is a temporary solution, which precludes
testing any translations in unit tests. Ideally we will be able to find
a way to redo our mocks in such a way that they are not triggered by the
inner workings of FastGettext.