(PUP-7114) Move semantic_puppet to lib/puppet/vendor
The vendored semantic_puppet ended up directly under /lib in the puppet
repository. The decision to put it there seems to be based on the
misunderstanding to otherwise be forced to change `require` statements.

Since that's not needed, this commit moves the semantic_puppet source
to lib/puppet/vendor to avoid future confusion.