(PUP-7102) Ensure that entries logged early during boot aren't lost
Some logic in Puppet, notably the loading of environments, happen very
early. So early in fact, that there is no logger configured. The `Log`
class will queue all entries logged at that time and then flush the
queue once it's initialized.

The problem is that if an exception is raised during this time, that
causes the process to terminate, those log entries are lost. With them,
valid information that led up to the exception are also lost.

This commit ensures that if queued log entries exists, and if there is
no logger configured, that those entries are flushed using a console
logger.