(PUP-7102) Log a warning when environment path has permission problems
Prior to this commit, a permission problem in one of the directories in
an `environmentpath` would go unnoticed. The directory would simply be
deemed invalid and excluded from the path. This would later lead to
an `EnvironmentNotFound` exception being raised with no clue as to why
the path wasn't included.

This commit ensures that the directory with permission problems is logged
as a warning (once). The warning can be silenced using a `disable_warnings`
setting that includes 'permission_denied'.