(PUP-3170) Add cap to Integer and forbid BigDecimal values
Before this, it was possible for Bignum values to escape Ruby into
Puppet where they could lead to strange effects when being stored in PDB
where values are truncated. BigDecimal values were also truncated.

The puppet language specifies the max size of the Puppet Integer and
Float types. These caps are now enforced as follows:

* When assigned to a resource parameter (at any depth in a structure)
* When passed in the 3x function API (at any depth in a structure)
* When serializing with Pcore serialiser (already in place - not added
in this commit).

After this commit, it is still allowed for 4.x functions to get and
return Bignum values as they are considered to be temporary result
integers. Performance would be degraded to much to add this. (In
comparison to the 3.x conversion that has to take place for other
reasons anyway).