(PUP-6006) Change phrasing of multiple errors final error
Before this, the error message for multiple errors simply said "multiple
errors. Giving up".
This commit changes this to say "Language valiation logged <n> errors.
Giving up". This tells the user that a) there is things in the log to
look at, and b) it was language validation (syntax errors and the like)
that caused the reported problems.