(pup-7156) add shim for semantic
In c111fb7, puppet switched to using the 'semantic_puppet' library instead of
'semantic'. Thus, the vendored libraries changed - 'semantic' was removed and
'semantic_puppet' added. Unexpectedly, this broke users of
garethr/puppet-module-skeleton, as this project at the time directly loaded the
'semantic' library via "require 'puppet/vendor/semantic/lib/semantic'".

This commit introduces a shim for puppet/vendor/semantic, so that users who
have required the library will not be broken. A warning is emitted as well,
noting that puppet's vendored libraries are private internal API and not
intended for external/public use.